requirements.txt
Flask
numpy
Pillow
scikit-learn
tensorflow
joblib
web3
pinatapy
